Separate out tests & env variables
