+++ packages/node/test/unit/node.spec.ts
@@ -1,28 +1,10 @@
 import * as ethers from "ethers";
-import firebase from "firebase";
-import firebasemock from "firebase-mock";
 
 import Node from "../../src/node";
-import {
-  Address,
-  AppInstanceInfo,
-  MethodName,
-  MethodRequest,
-  MethodResponse
-} from "../../src/node-types";
+
+import { firestore, networkContext, privateKey, provider } from "./env";
 
 describe("Node operations", () => {
-  const privateKey = process.env.npm_package_config_unlockedAccount0!;
-  const provider = ethers.getDefaultProvider() as ethers.providers.Web3Provider;
-  let firestore: firebase.firestore.Firestore;
-  let networkContext: Map<string, Address>;
-
-  beforeEach(() => {
-    firestore = new firebasemock.MockFirestore();
-    networkContext = new Map<string, Address>();
-    networkContext.set("AppRegistry", ethers.constants.AddressZero);
-  });
-
   it("exists", () => {
     expect(Node).toBeDefined();
   });
@@ -36,26 +18,5 @@ describe("Node operations", () => {
     const node = new Node(privateKey, provider, firestore, networkContext);
     const multisig = await node.createMultisig(ethers.constants.AddressZero);
     expect(multisig).toBeDefined();
-  });
-
-  it("can accept call to get app instances", done => {
-    const node = new Node(privateKey, provider, firestore, networkContext);
-    const requestId = "1";
-    const req: MethodRequest = {
-      requestId,
-      type: MethodName.GET_APP_INSTANCES,
-      params: {}
-    };
-
-    // Set up listener for the method response
-    node.on(req.type, (res: MethodResponse) => {
-      expect(res.result).toEqual({
-        appInstances: [] as AppInstanceInfo[]
-      });
-      done();
-    });
-
-    // Make the method call
-    node.emit(req.type, req);
   });
 });
